+++ src/index.ts
@@ -23,7 +23,7 @@ import KandelSeeder from "./kandel/kandelSeeder";
 import CoreKandelInstance from "./kandel/coreKandelInstance";
 import OfferMaker from "./offerMaker";
 import { enableLogging } from "./util/logger";
-import { addressesConfiguration, tokensConfiguration } from "./configuration";
+import configuration from "./configuration";
 import TickPriceHelper from "./util/tickPriceHelper";
 import GeometricKandelDistribution from "./kandel/geometricKandel/geometricKandelDistribution";
 import GeneralKandelDistribution from "./kandel/generalKandelDistribution";
@@ -56,6 +56,5 @@ export {
   OfferMaker,
   TickPriceHelper,
   enableLogging,
-  addressesConfiguration,
-  tokensConfiguration,
+  configuration,
 };